cxr-backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
torch>=2.0.0
torchvision>=0.15.0
torchxrayvision>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
python-multipart>=0.0.6
anthropic>=0.25.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
# Zero-shot CLIP model (CheXzero-style)
transformers>=4.35.0
